bl: use ld in hpm_sdk 1.8.0
