+++ projects/HSLink-Pro/bootloader/flash_xip.ld
@@ -7,15 +7,17 @@ ENTRY(_start)
 
 STACK_SIZE = _stack_size;
 HEAP_SIZE = _heap_size;
+/* BL给128K */
+_flash_size = 128K;
 
 /* 预留512字节的空间到AHB_SRAM给bootloader做配置 */
 __bootloader_reserved_length__ = 512;
 
 MEMORY
 {
-    XPI0 (rx) : ORIGIN = 0x80000000, LENGTH = 128k
+    XPI0 (rx) : ORIGIN = 0x80000000, LENGTH = _flash_size
     ILM (wx) : ORIGIN = 0x00000000, LENGTH = 128K
-    DLM (w) : ORIGIN = 0x00080000, LENGTH = 128K
+    DLM (w) : ORIGIN = 0x00080300, LENGTH = 128K - 768
     AHB_SRAM (w) : ORIGIN = 0xf0400000, LENGTH = 32K
 }
 
@@ -51,8 +53,10 @@ SECTIONS
         __vector_ram_start__ = .;
         KEEP(*(.vector_table))
         KEEP(*(.isr_vector))
+        KEEP(*(.isr_vector.*))
         KEEP(*(.vector_s_table))
         KEEP(*(.isr_s_vector))
+        KEEP(*(.isr_s_vector.*))
         . = ALIGN(8);
         __vector_ram_end__ = .;
     } > ILM
@@ -121,7 +125,6 @@ SECTIONS
     }  > XPI0
     __eh_frame_hdr_start = SIZEOF(.eh_frame_hdr) > 0 ? ADDR(.eh_frame_hdr) : 0;
     __eh_frame_hdr_end = SIZEOF(.eh_frame_hdr) > 0 ? . : 0;
-
 
     .rel : {
         KEEP(*(.rel*))
@@ -221,6 +224,7 @@ SECTIONS
         . = ALIGN(8);
         __noncacheable_init_start__ = .;
         KEEP(*(.noncacheable.init))
+        KEEP(*(.noncacheable.init.*))
         __noncacheable_init_end__ = .;
         . = ALIGN(8);
     } > DLM
@@ -230,6 +234,7 @@ SECTIONS
         . = ALIGN(8);
         __fast_ram_init_start__ = .;
         KEEP(*(.fast_ram.init))
+        KEEP(*(.fast_ram.init.*))
         __fast_ram_init_end__ = .;
         . = ALIGN(8);
     } > DLM
@@ -252,23 +257,30 @@ SECTIONS
     .framebuffer (NOLOAD) : {
         . = ALIGN(8);
         KEEP(*(.framebuffer))
+        KEEP(*(.framebuffer.*))
         . = ALIGN(8);
     } > DLM
 
-    .noncacheable.bss (NOLOAD) : {
+    .noncacheable (NOLOAD) : {
         . = ALIGN(8);
-        KEEP(*(.noncacheable))
+        KEEP(*(.noncacheable))  /* Legacy sections, kept for backwards compatibility */
+        KEEP(*(.noncacheable.non_init))
+        KEEP(*(.noncacheable.non_init.*))
         __noncacheable_bss_start__ = .;
         KEEP(*(.noncacheable.bss))
+        KEEP(*(.noncacheable.bss.*))
         __noncacheable_bss_end__ = .;
         . = ALIGN(8);
     } > DLM
 
-    .fast_ram.bss (NOLOAD) : {
+    .fast_ram (NOLOAD) : {
         . = ALIGN(8);
-        KEEP(*(.fast_ram))
+        KEEP(*(.fast_ram))  /* Legacy sections, kept for backwards compatibility */
+        KEEP(*(.fast_ram.non_init))
+        KEEP(*(.fast_ram.non_init.*))
         __fast_ram_bss_start__ = .;
         KEEP(*(.fast_ram.bss))
+        KEEP(*(.fast_ram.bss.*))
         __fast_ram_bss_end__ = .;
         . = ALIGN(8);
     } > DLM
@@ -291,6 +303,7 @@ SECTIONS
 
     .ahb_sram (NOLOAD) : {
         KEEP(*(.ahb_sram))
+        KEEP(*(.ahb_sram.*))
     } > AHB_SRAM
 
     __boot_header_load_addr__ = ORIGIN(AHB_SRAM) + LENGTH(AHB_SRAM) - __bootloader_reserved_length__;
