+++ projects/HSLink-Pro/bootloader/main.c
@@ -1,7 +1,7 @@
 #include "BL_Setting_Common.h"
 #include "bootuf2.h"
+#include "third_party_components/port/neopixel/ws2812.h"
 #include "usb_config.h"
-#include "ws2812.h"
 #include <board.h>
 #include <hpm_dma_mgr.h>
 #include <hpm_ewdg_drv.h>
